reorganizing into single project, since that is what I recall being the decision. Single project, single function app, multiple functions
