+++ fns-csv-pipeline/README.md
@@ -1,1 +1,1 @@
-CSV pipeline functions project: Receiver-Debatch, Validate, Transform.
+CSV pipeline functions project: Decompress, Validate, Transform.
