--- fns-csv-pipeline/fn-decompressor/.vscode/extensions.json
@@ -1,6 +1,0 @@
-{
-  "recommendations": [
-    "ms-azuretools.vscode-azurefunctions",
-    "vscjava.vscode-java-debug"
-  ]
-}
