--- fns-csv-pipeline/fn-decompressor/.vscode/launch.json
@@ -1,13 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Attach to Java Functions",
-            "type": "java",
-            "request": "attach",
-            "hostName": "127.0.0.1",
-            "port": 5005,
-            "preLaunchTask": "func: host start"
-        }
-    ]
-}
