--- fns-csv-pipeline/fn-decompressor/.vscode/settings.json
@@ -1,8 +1,0 @@
-{
-    "azureFunctions.javaBuildTool": "maven",
-    "azureFunctions.deploySubpath": "target/azure-functions/dex-csv-fn-app",
-    "azureFunctions.projectLanguage": "Java",
-    "azureFunctions.projectRuntime": "~4",
-    "debug.internalConsoleOptions": "neverOpen",
-    "azureFunctions.preDeployTask": "package (functions)"
-}
