--- fns-csv-pipeline/fn-decompressor/.vscode/tasks.json
@@ -1,25 +1,0 @@
-{
-	"version": "2.0.0",
-	"tasks": [
-		{
-			"type": "func",
-			"label": "func: host start",
-			"command": "host start",
-			"problemMatcher": "$func-java-watch",
-			"isBackground": true,
-			"options": {
-				"cwd": "${workspaceFolder}/target/azure-functions/dex-csv-fn-app"
-			},
-			"dependsOn": "package (functions)"
-		},
-		{
-			"label": "package (functions)",
-			"command": "mvn clean package",
-			"type": "shell",
-			"group": {
-				"kind": "build",
-				"isDefault": true
-			}
-		}
-	]
-}
