--- fns-csv-pipeline/fn-decompressor/README.md
@@ -1,2 +1,0 @@
-# Receiver-Debatcher Function for the CSV Pipeline
-this needs to be fleshed out, following the example of the HL7 pipeline
