+++ fns-csv-pipeline/pom.xml
@@ -3,11 +3,11 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>gov.cdc.dataexchange.csv</groupId>
-    <artifactId>decompressor</artifactId>
+    <artifactId>fns-csv-pipeline</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>DEX Fns-CSV-Pipeline :: decompressor</name>
+    <name>DEX Fns-CSV-Pipeline</name>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
