+++ fns-csv-pipeline/src/main/kotlin/gov/cdc/dex/csv/dtos/AzureBlobCreateEventMessage.kt
@@ -1,9 +1,9 @@
-package gov.cdc.dex.csv.decompressor
+package gov.cdc.dex.csv.dtos
 
 import com.google.gson.annotations.SerializedName
 
 //NOTE: there are more fields in the message, but these are the only ones we care about
-data class AzBlobCreateEventMessage (
+data class AzureBlobCreateEventMessage (
     @SerializedName("eventType" ) var eventType : String,
     @SerializedName("id"        ) var id        : String,
     @SerializedName("data"      ) var evHubData : EvHubData,
