+++ fns-csv-pipeline/src/main/kotlin/gov/cdc/dex/csv/services/AzureBlobServiceImpl.kt
@@ -1,0 +1,5 @@
+package gov.cdc.dex.csv.services
+
+class AzureBlobServiceImpl : BlobService{
+
+}
