+++ fns-csv-pipeline/src/main/kotlin/gov/cdc/dex/csv/services/BlobService.kt
@@ -1,0 +1,3 @@
+package gov.cdc.dex.csv.services
+
+interface BlobService{}
